Added total number of AffixTiers in parentheses T#(Total)
Setting ShowAffixBracketTierTotal=1 will enable display of the maximum
number of AffixTiers for a given Affix.

Off:
"+Armor.... T4 P"

On:
"+Armor.... T4(8) P"

I find this functionality useful in knowing just how good (or bad) an
item's tier is.

This uses the ShowAffixBracketTierTotal variable in Settings.ini.  I
have wired up the logic in the UI, but do not have a UI editor to add a
checkbox to the options dialog.  When ShowAffixBracketTier, a control
named ShowAffixBracketTierTotal will hide as you would expect.  All that
needs to be added is the actual visual control.